e2etest: move firewall from validator-tests
This commit moves interface of firewall actor from validator-test into
e2etest-firewall. It also updates crate references.
