+++ crates/validator/src/lib.rs
@@ -23,6 +23,7 @@ mod similarity_functions;
 use async_backtrace::framed;
 use e2etest_dns::Dns;
 use e2etest_dns::DnsExt;
+use e2etest_firewall::Firewall;
 use std::net::Ipv4Addr;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -33,7 +34,6 @@ use tracing_subscriber::EnvFilter;
 use tracing_subscriber::filter;
 use tracing_subscriber::fmt;
 use tracing_subscriber::prelude::*;
-use vector_search_validator_tests::Firewall;
 use vector_search_validator_tests::ScyllaCluster;
 use vector_search_validator_tests::ScyllaClusterExt;
 use vector_search_validator_tests::ScyllaProxyCluster;
